requirements.txt
pyaes==1.6.1
Pyrogram==2.0.100
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
TgCrypto==1.2.5
